график анализа вынесен в пакет
