+++ local_libs/pie_chart_package/lib/pie_chart_package.dart
@@ -1,0 +1,2 @@
+export 'src/models/pie_chart_data_model.dart';
+export 'src/animated_pie_chart.dart';
