+++ local_libs/pie_chart_package/lib/src/models/pie_chart_data_model.dart
@@ -1,0 +1,11 @@
+class PieChartDataModel {
+  final String id;
+  final String name;
+  final double percent;
+
+  const PieChartDataModel({
+    required this.id,
+    required this.name,
+    required this.percent,
+  });
+}
