+++ pubspec.yaml
@@ -29,8 +29,10 @@ dependencies:
   fuzzy: ^0.5.1
   sqflite: ^2.4.2
   path: ^1.9.1 # надо для sqflite
-  fl_chart: ^1.0.0
   flutter_launcher_icons: ^0.14.4
+
+  pie_chart_package:
+    path: ./local_libs/pie_chart_package
 
 dev_dependencies:
   flutter_test:
